Properly support non-default namespaces
I noticed that some but not all pods from releases installed in
a namespace other than "default" do not show up in Capacitor.

It turns out that if a release is installed a non-defult namespace
Capacitor would incorrectly detect the corresponding artifacts as
belonging to the "default" namespace unless the chart templates
explicitly set the "metadata.namespace" attribtue even though
that is not required.

Installing the "podinfo" chart in a non-default namespace is a good
way to reproduce this.